+++ tests/pendulum_tests/test_fluent_setters.py
@@ -87,6 +87,17 @@ class FluentSettersTest(AbstractTestCase):
         self.assertEqual(7, d.month)
         self.assertEqual(2, d.day)
 
+    def test_fluid_on(self):
+        d = Pendulum.create(2016, 7, 2, 0, 41, 20)
+        new = d.on(1995, 11, 9)
+        self.assertIsInstanceOfPendulum(new)
+        self.assertEqual(1995, new.year)
+        self.assertEqual(11, new.month)
+        self.assertEqual(9, new.day)
+        self.assertEqual(2016, d.year)
+        self.assertEqual(7, d.month)
+        self.assertEqual(2, d.day)
+
     def test_fluid_with_date_with_transition(self):
         d = Pendulum.create(2013, 3, 31, 0, 0, 0, 0, 'Europe/Paris')
         new = d.with_date(2013, 4, 1)
@@ -100,9 +111,33 @@ class FluentSettersTest(AbstractTestCase):
         self.assertEqual(31, d.day)
         self.assertEqual(3600, d.offset)
 
-    def test_fluid_set_time(self):
+    def test_fluid_on_with_transition(self):
+        d = Pendulum.create(2013, 3, 31, 0, 0, 0, 0, 'Europe/Paris')
+        new = d.on(2013, 4, 1)
+        self.assertIsInstanceOfPendulum(new)
+        self.assertEqual(2013, new.year)
+        self.assertEqual(4, new.month)
+        self.assertEqual(1, new.day)
+        self.assertEqual(7200, new.offset)
+        self.assertEqual(2013, d.year)
+        self.assertEqual(3, d.month)
+        self.assertEqual(31, d.day)
+        self.assertEqual(3600, d.offset)
+
+    def test_fluid_with_time(self):
         d = Pendulum.create(2016, 7, 2, 0, 41, 20)
         new = d.with_time(5, 32, 49)
+        self.assertIsInstanceOfPendulum(new)
+        self.assertEqual(5, new.hour)
+        self.assertEqual(32, new.minute)
+        self.assertEqual(49, new.second)
+        self.assertEqual(0, d.hour)
+        self.assertEqual(41, d.minute)
+        self.assertEqual(20, d.second)
+
+    def test_fluid_at(self):
+        d = Pendulum.create(2016, 7, 2, 0, 41, 20)
+        new = d.at(5, 32, 49)
         self.assertIsInstanceOfPendulum(new)
         self.assertEqual(5, new.hour)
         self.assertEqual(32, new.minute)
@@ -114,6 +149,14 @@ class FluentSettersTest(AbstractTestCase):
     def test_fluid_set_time_with_transition(self):
         d = Pendulum.create(2013, 3, 31, 0, 0, 0, 0, 'Europe/Paris')
         new = d.with_time(2, 30, 0)
+        self.assertIsInstanceOfPendulum(new)
+        self.assertEqual(3, new.hour)
+        self.assertEqual(30, new.minute)
+        self.assertEqual(0, new.second)
+
+    def test_fluid_at_with_transition(self):
+        d = Pendulum.create(2013, 3, 31, 0, 0, 0, 0, 'Europe/Paris')
+        new = d.at(2, 30, 0)
         self.assertIsInstanceOfPendulum(new)
         self.assertEqual(3, new.hour)
         self.assertEqual(30, new.minute)